# docs/swagger.yaml
basePath: /
definitions:
...
  models.DiscountCashDto:
    properties:
      balance_discount_cash:
        type: number
      balance_discount_near_expire:
        type: number
      discount_cash_near_expire_date:
        type: string
      discounts:
        items:
          $ref: '#/definitions/models.DiscountsDto'
        type: array
      member_tel:
        type: string
    type: object
  models.DiscountsDto:
    properties:
      balance_discount_cash:
        type: number
      deposit_id:
        type: string
      discount_cash_expire_date:
        type: string
    type: object
...
  models.GroupMenuDto:
    properties:
      display_name:
        type: string
      enable_member:
        type: boolean
      is_active:
        type: boolean
      name:
        type: string
      receipt_stub:
        type: boolean
      sorting:
        type: integer
    type: object
...
  utils.StandardErrorResponse:
    properties:
      data:
        description: "null"
      message:
        example: Failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.StandardSuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: This is a POS API with Gin and Swagger
  title: New POS APIls
  version: "1.0"
paths:
...
  /api/customer-tier/{tel}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Tel
        in: path
        name: tel
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get customer tier by tel
      tags:
      - Customer Tier
  /api/customer-tier/import:
    post:
      consumes:
      - multipart/form-data
      description: Import customer tier from CSV or XLSX
      parameters:
      - description: CSV or XLSX file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Import customer tier
      tags:
      - Customer Tier
  /api/customer-tier/import/first:
    post:
      consumes:
      - multipart/form-data
      description: Import customer tier from CSV or XLSX
      parameters:
      - description: CSV or XLSX file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Import customer tier first time
      tags:
      - Customer Tier
  /api/deposit-cron/{tel}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Telephone number
        in: path
        name: tel
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get deposit head cron by tel
      tags:
      - Deposit Head
  /api/discountcash/:member_tel:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DiscountCashDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      tags:
      - Discount Cash
  /api/group-menu:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get group menu all active
      tags:
      - Group Menu
    post:
      consumes:
      - application/json
      parameters:
      - description: Group Menu Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GroupMenuDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a group menu
      tags:
      - Group Menu
  /api/group-menu/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete group menu by ID
      tags:
      - Group Menu
    get:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get group menu by id
      tags:
      - Group Menu
    put:
      consumes:
      - application/json
      parameters:
      - description: Group Menu Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GroupMenuDto'
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update group menu by ID
      tags:
      - Group Menu
  /api/group-menu/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search text
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get group menu list with paggination
      tags:
      - Group Menu
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"